% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_total_time}
\alias{get_total_time}
\title{Total time}
\usage{
get_total_time(tab)
}
\arguments{
\item{tab}{dataframe with two columns "PIN" and "timestamp"}
}
\value{
returns dataframe with two columns: "PIN" and "time". "PIN" is ids of participants, "Time" is amount of time spent on the game in seconds
}
\description{
function computes total amount of time spent interacting with the app by subtracting timestamp of last entry from timestamp of each PIN
}
