% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_total_versions}
\alias{get_total_versions}
\title{Versions of the game in the batch}
\usage{
get_total_versions(tab)
}
\arguments{
\item{tab}{dataframe with two columns "PIN" and "Version"}
}
\value{
returns dataframe with two columns: "PIN" and "version". "PIN" is ids of participants, "Version" is version of the game
}
\description{
function extracts all versions of the game that appeared in the current batch
}
